[chores] Added mixin for admin inline tests #166
To reduce duplicacy, added a mixin to abstract the common
methods required for testing inline admin.
Added fallback for cases where load event is fired earlier.
Added explicit link to selenium mixin in README.
Added common CI tasks.

Fixes #166
