+++ .github/workflows/ci.yml
@@ -92,6 +92,11 @@ jobs:
         coverage xml
       env:
         SELENIUM_HEADLESS: 1
+        GECKO_LOG: 1
+
+    - name: Show gecko web driver log on failures
+      if: ${{ failure() }}
+      run: cat geckodriver.log
 
     - name: Upload Coverage
       if: ${{ success() }}
