+++ README.rst
@@ -544,13 +544,15 @@ Launch development server and SMTP debugging server:
 
 You can access the admin interface at http://127.0.0.1:8000/admin/.
 
-Run tests with (make sure you have the :ref:`selenium dependencies
-<selenium_dependencies>` installed locally first):
+Run tests with:
 
 .. code-block:: shell
 
     # pytests is used to test django-channels
     ./runtests.py && pytest
+
+Selenium dependencies are included with `openwisp-utils
+<https://openwisp.io/docs/dev/utils/developer/test-utilities.html#openwisp-utils-tests-seleniumtestmixin>`_.
 
 Contributing
 ------------
